Decomposed GameService
