+++ packages/api/api/handlers/warp_well.py
@@ -1,7 +1,6 @@
 """Shared turn-scoped warp-well route logic for Core and BFF callers."""
 
-from api.concepts.warp_well import WarpWellKind
-from api.services.game_service import GameService
+from api.services.turn_concept_service import TurnConceptService
 from api.transport.concept_warp_well import (
     CoordinateInWarpWellRequest,
     CoordinateInWarpWellResponse,
@@ -11,14 +10,15 @@ from api.transport.concept_warp_well import (
 
 
 def coordinate_in_well(
-    svc: GameService,
+    svc: TurnConceptService,
     game_id: int,
     perspective: int,
     turn_number: int,
     body: CoordinateInWarpWellRequest,
 ) -> CoordinateInWarpWellResponse:
     """Return whether ``(map_x, map_y)`` lies in the given warp well of the planet."""
-    kind = WarpWellKind(body.well_type.value)
+    from api.concepts.warp_well import WarpWellKind
+
     inside = svc.warp_well_coordinate_in_well(
         game_id,
         perspective,
@@ -26,13 +26,13 @@ def coordinate_in_well(
         body.planet_id,
         body.map_x,
         body.map_y,
-        kind,
+        WarpWellKind(body.well_type.value),
     )
     return CoordinateInWarpWellResponse(inside=inside)
 
 
 def warp_well_cells(
-    svc: GameService,
+    svc: TurnConceptService,
     game_id: int,
     perspective: int,
     turn_number: int,
@@ -40,6 +40,13 @@ def warp_well_cells(
     well_type: WarpWellTypeParam,
 ) -> WarpWellCellsResponse:
     """Return map cell indices whose centers lie in the given warp well."""
-    kind = WarpWellKind(well_type.value)
-    cells = svc.warp_well_cells(game_id, perspective, turn_number, planet_id, kind)
+    from api.concepts.warp_well import WarpWellKind
+
+    cells = svc.warp_well_cells(
+        game_id,
+        perspective,
+        turn_number,
+        planet_id,
+        WarpWellKind(well_type.value),
+    )
     return WarpWellCellsResponse(cells=cells)
