+++ packages/api/tests/test_warp_well_consistency.py
@@ -10,7 +10,7 @@ from api.concepts.warp_well import (
     coordinate_in_warp_well,
     map_cell_indices_in_warp_well,
 )
-from api.services.game_service import GameService
+from api.services.stack import build_service_stack
 from api.storage.memory_asset import MemoryAssetBackend
 
 FIXTURE_PATH = Path(__file__).resolve().parents[3] / "test-fixtures" / "warp-well-consistency.json"
@@ -24,8 +24,8 @@ def template_planet():
         backend.put("games/628580/info", json.load(f))
     with open(ASSETS_DIR / "turn_sample.json") as f:
         backend.put("games/628580/1/turns/111", json.load(f))
-    svc = GameService(backend)
-    return svc.get_planet_from_turn(628580, 1, 111, 1)
+    _, turns, _, _ = build_service_stack(backend)
+    return turns.get_planet_from_turn(628580, 1, 111, 1)
 
 
 @pytest.fixture(scope="module")
